+++ src/integrationTest/java/com/kryptokrauts/aeternity/test/integration/BaseTest.java
@@ -35,6 +35,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
+import java.util.stream.Collectors;
 import javax.naming.ConfigurationException;
 import org.apache.commons.io.IOUtils;
 import org.junit.After;
@@ -69,6 +70,19 @@ public abstract class BaseTest {
 
   private static final String MDW_BASE_URL = "MDW_BASE_URL";
 
+  private static final List<String> testPrivateKeys =
+      List.of(
+          "7c6e602a94f30e4ea7edabe4376314f69ba7eaa2f355ecedb339df847b6f0d80575f81ffb0a297b7725dc671da0b1769b1fc5cbe45385c7b5ad1fc2eaf1d609d",
+          "7fa7934d142c8c1c944e1585ec700f671cbc71fb035dc9e54ee4fb880edfe8d974f58feba752ae0426ecbee3a31414d8e6b3335d64ec416f3e574e106c7e5412",
+          "1509d7d0e113528528b7ce4bf72c3a027bcc98656e46ceafcfa63e56597ec0d8206ff07f99ea517b7a028da8884fb399a2e3f85792fe418966991ba09b192c91",
+          "58bd39ded1e3907f0b9c1fbaa4456493519995d524d168e0b04e86400f4aa13937bcec56026494dcf9b19061559255d78deea3281ac649ca307ead34346fa621",
+          "50458d629ae7109a98e098c51c29ec39c9aea9444526692b1924660b5e2309c7c55aeddd5ebddbd4c6970e91f56e8aaa04eb52a1224c6c783196802e136b9459",
+          "707881878eacacce4db463de9c7bf858b95c3144d52fafed4a41ffd666597d0393d23cf31fcd12324cd45d4784d08953e8df8283d129f357463e6795b40e88aa",
+          "9262701814da8149615d025377e2a08b5f10a6d33d1acaf2f5e703e87fe19c83569ecc7803d297fde01758f1bdc9e0c2eb666865284dff8fa39edb2267de70db",
+          "e15908673cda8a171ea31333538437460d9ca1d8ba2e61c31a9a3d01a8158c398a14cd12266e480f85cc1dc3239ed5cfa99f3d6955082446bebfe961449dc48b",
+          "6eb127925aa10d6d468630a0ca28ff5e1b8ad00db151fdcc4878362514d6ae865951b78cf5ef047cab42218e0d5a4020ad34821ca043c0f1febd27aaa87d5ed7",
+          "36595b50bf097cd19423c40ee66b117ed15fc5ec03d8676796bdf32bc8fe367d82517293a0f82362eb4f93d0de77af5724fba64cbcf55542328bc173dbe13d33");
+
   protected static final VirtualMachine targetVM = VirtualMachine.FATE;
 
   protected KeyPairService keyPairService;
@@ -77,12 +91,15 @@ public abstract class BaseTest {
 
   protected KeyPair keyPair;
 
+  protected List<KeyPair> otherKeyPairs;
+
   protected static String paymentSplitterSource,
       ethereumSignaturesSource,
       chatBotSource,
       gaBlindAuthSource,
       sophiaTypesSource,
-      ecdsaAuthSource;
+      ecdsaAuthSource,
+      statefulTest;
 
   protected UnitConversionService unitConversionService = new DefaultUnitConversionServiceImpl();
 
@@ -109,6 +126,11 @@ public abstract class BaseTest {
     keyPairService = new KeyPairServiceFactory().getService();
 
     keyPair = keyPairService.recoverKeyPair(TestConstants.BENEFICIARY_PRIVATE_KEY);
+
+    otherKeyPairs =
+        testPrivateKeys.stream()
+            .map(key -> keyPairService.recoverKeyPair(key))
+            .collect(Collectors.toList());
 
     aeternityService =
         new AeternityServiceFactory()
@@ -174,6 +196,7 @@ public abstract class BaseTest {
     gaBlindAuthSource = getContractSourceCode("GaBlindAuth.aes");
     sophiaTypesSource = getContractSourceCode("SophiaTypes.aes");
     ecdsaAuthSource = getContractSourceCode("ECDSAAuth.aes");
+    statefulTest = getContractSourceCode("StatefulTest.aes");
   }
 
   protected BigInteger getNextKeypairNonce() {